requirements.txt
backcall==0.1.0
cachetools==3.1.0
decorator==4.4.0
fire==0.1.3
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
oauth2client==4.1.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyDrive==1.3.1
Pygments==2.4.0
python-dotenv==0.10.2
PyYAML==5.1
rsa==4.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
virtualenv==16.5.0
wcwidth==0.1.7
